Move item type enum to new folder
