+++ src/main/kotlin/io/github/soat7/myburguercontrol/domain/model/Item.kt
@@ -1,7 +1,7 @@
 package io.github.soat7.myburguercontrol.domain.model
 
+import io.github.soat7.myburguercontrol.domain.enum.ItemType
 import java.math.BigDecimal
-import java.util.Locale
 import java.util.UUID
 
 data class Item(
@@ -10,21 +10,3 @@ data class Item(
     val price: BigDecimal,
     val type: ItemType
 )
-
-enum class ItemType {
-    APPETIZER,
-    DESSERT,
-    DRINK,
-    FOOD,
-    OTHER;
-
-    companion object {
-        fun from(resource: String): ItemType {
-            return try {
-                ItemType.valueOf(resource.uppercase(Locale.getDefault()))
-            } catch (ex: IllegalArgumentException) {
-                throw ex
-            }
-        }
-    }
-}
