+++ src/main/kotlin/io/github/soat7/myburguercontrol/infrastructure/rest/item/api/ItemCreationRequest.kt
@@ -1,5 +1,6 @@
 package io.github.soat7.myburguercontrol.infrastructure.rest.item.api
 
+import io.github.soat7.myburguercontrol.domain.enum.ItemType
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
 import java.math.BigDecimal
@@ -11,13 +12,6 @@ data class ItemCreationRequest(
     @NotNull
     val price: BigDecimal,
 
+    @NotBlank
     val type: ItemType
-) {
-    enum class ItemType {
-        APPETIZER,
-        DESSERT,
-        DRINK,
-        FOOD,
-        OTHER
-    }
-}
+)
